requirements.txt
txt

requests==2.31.0
google-api-python-client==2.108.0
python-dotenv==1.0.0

